test(cli): split app dev server tests
